pyproject.toml
[tool.black]
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.vscode
  | \.hg
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output="3"
include_trailing_comma="True"
force_grid_wrap=0
use_parentheses="True"
line_length="88"

[tool.coverage.run]
omit = ["**/*__init__.py", "**/*__main__.py", "**/*_test.py", "**/logger.py"]

[tool.coverage.report]
fail_under = 20  # Force tests to fail if coverage is under the specified percentage

[tool.coverage.html]
directory = 'coverage.html'

[tool.poetry]
name = "climate_dataset"
version = "1.0.7"
description = ""
authors = ["Kushagra"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
Flask = "^1.1.2"
gunicorn = "^20.0.4"
python-dotenv = "^0.15.0"
flask-redoc = "^0.2.1"
debugpy = "^1.2.1"
google-api-core = "^2.2.2"
google-api-python-client = "^2.31.0"
google-auth = "^2.3.3"
google-cloud-pubsub = "^2.9.0"
google-cloud = "^0.34.0"
backoff = "^1.11.1"
pandas = "^1.3.4"
xlrd = "^2.0.1"
openpyxl = "^3.0.9"
google-cloud-storage = "^2.3.0"
gcsfs = "^2022.5.0"
geopy = "^2.2.0"
numpy = "^1.21.4"

[tool.poetry.dev-dependencies]
whippet = "^0.3.2"
pytest = "^6.2.1"
coverage = {extras = ["toml"], version = "^5.0.4"}
pytest-cov = "^2.11.1"
black = "^20.8b1"
bandit = "^1.7.0"
flake8 = "^3.8.4"
flake8-isort = "^4.0.0"
flake8-bugbear = "^20.11.1"
flake8-builtins = "^1.5.3"
flake8-2020 = "^1.6.0"
flake8-pep3101 = "^1.3.0"
flake8-variables-names = "^0.0.4"
flake8-AAA = "^0.9.0"
flake8-pytest = "^1.3"
pep8-naming = "^0.11.1"

[build-system]
requires = ["poetry==1.1.4"] # Same version as in Dockerfile
build-backend = "poetry.masonry.api"
